New damage.h, deleted from monster and hero
