+++ Monster.h
@@ -1,44 +1,14 @@
 #pragma once
 class Hero;
 #include "Hero.h"
+#include "Damage.h"
 #include <string>
 
 class Monster
 {
     protected:
-        struct Damage {
-
-            int physical;
-
-            int magical;
-
-            Damage operator+(const Damage& other){
-                    Damage dmgToReturn;
-                    dmgToReturn.physical = this->physical + other.physical;
-                    dmgToReturn.magical = this->magical + other.magical;
-
-                    return dmgToReturn;
-        }
-
-             Damage& operator+=(const Damage& other){
-                this->physical += other.physical;
-                this->magical += other.magical;
-                return *this;
-             }
-
-            Damage& operator*=(const Damage& other){
-                this->physical *= other.physical;
-                this->magical *= other.magical;
-                return *this;
-            }
-
-            friend std::ostream& operator<<(std::ostream& out, const Damage& dmg){
-                out<<"Physical: "<<dmg.physical<<", Magical: "<<dmg.magical;
-                return out;
-            }
-
-
-        }; Damage damage= {0,0};
+       
+        Damage damage= {0,0};
         int hp; ///< the hp of the Unit
         int dmg; ///< the damage of the Unit
         const std::string name; ///<The name of the Unit
